neomat_care/requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
dj-database-url==3.1.2
Django==6.0.1
django-csp==4.0
django-extensions==4.1
django-filter==25.2
django-sslserver==0.22
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
inflection==0.5.1
MarkupSafe==3.0.3
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
Werkzeug==3.1.5
